Fix axios POST requests
By default, axios posts data as JSON, not form-encoded. Using
URLSearchParams is enough to convince it to use form encoding instead
(we do the same thing for fetch, too).

In the tests, we use purge as an API action that requires POST without
actually being destructive or having any effects on the wiki.