Fix fetch POST requests
This requires a major change to how the requests are made. Previously,
we put all parameters in the POST request body, but the origin parameter
(usually origin=*) must actually be in the URL, so that it becomes part
of the CORS preflight request. Since this distinction isn’t really
limited to fetch, and it’s perfectly fine to send the origin parameter
(if there is any) in the URL with axios requests, we implement this in
core, not fetch: core now splits the parameters into two sets (for now,
only origin needs to go into the URL, but maybe others will be added
later), and internalPost takes two parameters, URL and body params.