Add cookie jar support to axios
This uses an in-memory cookie jar for the lifetime of the session, which
should be good enough for now. Maybe I’ll add support for other cookie
stores later.

The axiosCookieJarSupport module is a bit odd – apparently its default
export is an object with a default member which has the actual function,
hence the axiosCookieJarSupport.default( ... ) call. Maybe a future
version will improve on that.