Test that README example works
I want to ensure that I don’t accidentally break the README.md example,
so let’s add a script which pipes the first Markdown block into Node.js.

In order to verify that the script works, any promise failures need to
turn into a nonzero exit code. One way to ensure this would be to turn
the .catch( console.error ) into a block like

    .catch( e => {
        console.error( e );
	process.exitCode = 1;
    } );

but I felt that this would’ve bloated the example too much. The
alternative is to use top-level await; between Node 13.3 and 14.8, this
requires a feature flag (--harmony-top-level-await), but thankfully,
even the latest version of node (16.7.0) still accepts this flag without
complaining (it’s just ignored), so we can just add that flag to the
test script, and use top-level await without having to raise our
required Node version to 14.8 (where it became enabled by default).

As a bonus, this change saves an indentation level in the code, which is
useful because the tab width on both GitHub and npmjs.com is eight
characters, so indentation levels add up quickly.