Automatically retry requests (once, by default)
If the response contains a Retry-After header (which MediaWiki usually
sends if the request specified the maxlag parameter and the current lag
is above the parameter value), then m3api can now automatically wait for
the specified time and send the request again. By default, it does this
once, but this can be increased or disabled using the maxRetries request
option.

I’m not yet sure if this is a good idea; I think it’s a useful feature
to have, but maybe it shouldn’t be enabled by default, or maybe there
should be some cap on the delay in case MediaWiki responds with an
unreasonably high Retry-After value. But let’s try this out for now and
hopefully see how it works in practice.