Introduce CombiningSession, mix it into BrowserSession, NodeSession
This introduces a new Session subclass that automatically combines
compatible requests that are made within the same event loop iteration.
The class is also mixed into the BrowserSession and NodeSession classes,
so that its functionality becomes available automatically.

So far, requests only count as compatible if they have identical values
for a parameter or if only one of them specifies the parameter at all. A
future commit will add support for Sets, where the parameter values of
multiple requests will be merged. (Arrays, on the other hand, are never
merged, so the expectation is that most multi-valued parameters should
be specified as sets instead of arrays.)

To make eslint accept expressions like expect( ... ).not.to.be.empty,
which looks like a no-op with no side effects, we set up the plugin
eslint-plugin-chai-friendly.

Part of #2.