Support set parameters (can be combined)
This is the big missing piece of #2. Any number of concurrent compatible
API requests can now be combined automatically; see the “more than two
requests” tests for a somewhat contrived example. Two brief mentions of
set parameters are added to the README.md, but I’ll rework it more
generally in a separate commit.