Don’t combine incompatible PageSet/continuation params
generator/continue and titles/pageids/revids can’t be combined usefully;
with generator, titles/etc. become the input of the generator (which may
use them, e.g. generator=links, or ignore them), while with continue=
the titles/etc. may be ignored (I assume they’re expected to have been
returned in the first, non-continue response). See also comments on #2.

In a way, this feels uncomfortably non-minimal (in that it encodes quite
a bit of knowledge about API parameters), but it’s not specific to a
single API module (it’s mainly about action=query, but e.g. action=purge
also has the same parameters), so I can live with it.