Transform set elements when combining
If one request specifies a number as a number, and another specifies the
same number as a string, we shouldn’t send the number twice.

(Note that we don’t transform set elements if they never get combined:
in that case, it’s the caller’s fault, I would say.)

While we’re at it, also expand a related test case. (We don’t test other
scalar types within sets because that’s not supposed to be supported: a
set of boolean-like parameters wouldn’t make sense.)