Turn third constructor parameter into defaultOptions
This makes it more consistent with how the params/defaultParams
parameters of request() and constructor() work, and also increases
flexibility, both for the current options (e.g. you can create a session
that will never retry requests) as well as potential future expansions
(there might be other options that will make more sense as per-instance
defaults than per-request).

We temporarily still recognize the old syntax and turn a string into an
object with userAgent, but this isn’t even documented in CHANGES.md, and
will be removed without further notice in the release following the next
one.

Fixes #3.