Require warn option to be present
This is mainly relevant for the unit tests, which previously often
extended Session directly and didn’t provide a warn() handler or a user
agent; once we start warning about missing user agents (#4), that
doesn’t work well. Introduce a base class for unit tests with a default
warning and user agent.