Warn about default user agent once per session
If a request doesn’t specify a custom user agent, send a warning to its
warning handler. (This is allowed, since the documentation says the
handler can be called with Error instances, including but not limited to
ApiWarnings.) To reduce logspam, only do this once per session. The
relevant field is only added to the session instance after we’ve sent
the warning, so well-behaved users should never have to pay the few
extra bytes for their session instances, since they’ll never hit that
code path.

Fixes #4.