Use more switch/case in tests
I’ve started using this pattern instead of the const/if in m3api-query,
and I think I like it better. IIRC I was initially wary about it because
switch doesn’t check strict equality, but that seems overly paranoid /
dogmatic to me now: we know that the ++ operator must produce a number.