Don’t allow callers to modify continuation
This was an oversight that I noticed now: by changing the contents of
the `continue` object, callers could change how continuation proceeds.
While this could possibly be conceived of as a feature, I think it is
more likely to be due to a bug, so let’s prevent it; if anyone really
wants to do this, they can implement continuation themselves, there’s
not that much code in requestAndContinue().

Integrate a check for this into one of the existing requestAndContinue
tests, because a separate test case feels like way too much extra code.

Document it as a breaking change, because theoretically it is, though
post-1.0 I probably wouldn’t consider this worth a major version bump.