Drop truncatedresult warnings in requestAndContinueReducingBatch
Since we’re fully in control of the execution flow within a batch (from
which the user can only escape by throwing in the reducer), we know that
we’ll also retrieve the rest of the truncated result, and the warning
about incomplete data would only apply if the reducer was programmed to
drop the data from the following responses, despite this function being
the one you use when you don’t want that problem. In other words, this
function is the solution to the truncated responses problem (among
others), so we should hide the truncated responses warning.

We can reliably recognize the warning even in errorformat=bc (I think),
because in that case the message is always in English, and does not use
local (on-wiki) messages either: it’s effectively a string hard-coded
into MediaWiki. And within MediaWiki, the message has been very stable:
it accidentally gained an extra space between MediaWiki 1.26 and 1.28,
but otherwise it’s remained constant. So in cases where the warning
doesn’t have a code, we can check the message instead.

Finishes #14; see also lucaswerkmeister/m3api-query#2.