Update eslint and eslint-config-wikimedia
This pulls in the no-shadow rule, which I don’t like, so let’s disable
it; a mysterious es/no-promise-any rule that doesn’t actually work
(“Definition for rule 'es/no-promise-any' was not found”), so let’s
disable that too; and, inexplicably, comma-dangle linting on JSON files
(i.e. thousands of errors in package-lock.json), fixing which would
actually break the JSON syntax, so let’s disable that as well.

On the bright side, jsdoc/require-yields points out that the two
continuation methods should use @yield, not @return.