Introduce dropTruncatedResultWarning option
This makes this functionality, previously hard-coded in
requestAndContinueReducingBatch(), available in request() and
requestAndContinue() as well. The need for this became clear in
m3api-query – specifically, a queryFullRevisionByRevisionId() function
should not have to wait for the whole batch to complete once the
revision was found, since the API always returns revisions in full.