Add tokenType, tokenName options
With this, m3api gains the ability to automatically retrieve, store and
reuse CSRF tokens (see the integration test for usage). In rare cases
where (due to combined requests) the first response doesn’t include the
token yet, we follow continuation; note that this requires T309921 [1]
in MediaWiki (part of the upcoming 1.39 LTS release).

Part of #6; however, we don’t clear the tokens on badtoken errors yet,
so currently users have to do that manually.

[1]: https://phabricator.wikimedia.org/T309921