Handle badtoken errors
Clear all the cached tokens and then retry with no delay.

Note that we still clear the tokens explicitly in the integration test,
because we expect users to do this after logging in – the plan is to
make this.tokens public with #21, but also to offer separate packages
that log in and then clear tokens automatically.

Part of #6.