Consider default options compatible for combining
If one request sets an option to the default value (e.g. explicit
method: 'GET'), allow it to be combined with another request that
doesn’t set the option. (Do *not* allow this for non-default options,
though!) This is needed so that token requests can be combined with
other requests (and therefore part of #6).