Export DEFAULT_OPTIONS from core.js to combine.js
This is needed so that combine.js can learn to combine requests that
explicitly specify default options with ones that leave those options
unset, which in turn will be needed for tokens support (#6).

The DEFAULT_OPTIONS are documented as private for now; they’ll become
part of the stable interface with #21.