Combine requests with different dropTruncatedResultWarning
This is required for tokens support (#6), where we want the token
requests to drop these warnings (we follow continuation), but also want
to let the token requests be combined with other requests that don’t
specify this option. To implement it, bring the “drop warnings” code
closer to how it was prior to commit 2138581290, with much of the
makeWarnDroppingTruncatedResultWarning() function lifted from there, and
export this decorator so that combine.js can implement the option
dropTruncatedResultWarning on its own (and always pass `false` to the
lower layer).

I’m not documenting makeWarnDroppingTruncatedResultWarning() as private,
so in theory it’s part of the stable interface; I don’t recommend using
it though: the request option is usually more convenient.