Rename isTruncatedResultWarning() to notTruncatedResultWarning()
I realized while working on the previous commit that this function was
actually named incorrectly: it returns true if the given warning *isn’t*
a truncatedresult warning. Since the function is easier to use this way,
rename the function while keeping the implementation the same. (It’s not
exported or a part of the stable interface anyways.)