Move warning handling into internalRequest()
This is needed for token handling (#6), where we will need to call a
request()-like method (getToken(), with an options parameter) from
internalRequest(), so internalRequest() needs a warn option. (The
request for the token must happen in internalRequest(), not request(),
because the token needs to be re-requested after a badtoken error, and I
don’t want to make internalRequest() call request() on retry.)

This also seems like a sensible follow-up to the previous commit, which
moved error handling into internalRequest(). We’re also again replacing
a private-but-reachable method with an unexported function.