Add retryAfterMaxlagSeconds, retryAfterReadonlySeconds options
We still retry whenever we get a Retry-After response header, but now we
can also retry if we *don’t* get the header, as long as the error is
either maxlag or readonly; the default header value is supplied by these
two options in that case.

For maxlag, this is rarely expected to be relevant, since MediaWiki
sends the header, and we only use the option if the header is absent.
For example, if the header is 10, the option is 5, and a 10-second retry
would no longer be within the maxRetriesSeconds, but a 5-second one
would be, then we still don’t retry after 5 seconds. The option is
mainly included for consistency (a real-world scenario where it might be
used would be a badly configured proxy dropping the header, I guess).

For readonly errors, this option is much more relevant, since MediaWiki
doesn’t (yet?) send this response header for those errors, but we still
want to automatically retry requests there (with a longer delay).

Part of #1.