Replace maxRetries with maxRetriesSeconds
Instead of limiting the *number* of retries, limit their *duration* (not
just of the delay itself but also of the repeated network requests – but
excluding the final request); the longer default of 65 seconds should be
enough to cover a brief read-only period caused by replication lag or DB
maintenance. (Though MediaWiki doesn’t send Retry-After for readonly API
errors, and we don’t have a default for that yet – will be added later.)

Part of #1.