Allow token requests to be combined
When a request with a tokenType comes in, eagerly prefetch the token, so
that *that* request can be combined with other query requests happening
simultaneously. As part of that, also treat requests with different
tokenName params as compatible if the tokenType is null (because the
tokenName has no effect in that case) – when core.js requests the token,
it clears the tokenType option but leaves the tokenName untouched, so if
we didn’t ignore the tokenName in that case, the token request for a
non-default tokenName would be incompatible with other requests.

Part of #6.