Fix running browser tests in CI
Apparently mocha-headless doesn’t work with newer Chrome/Chromium
versions – it broke on my local system a while ago, and eventually also
in CI. I haven’t been able to find a working replacement (Karma looked
most promising, but seems to be married to RequireJS too tightly to
support modules as test files), and in the meantime Node released
experimental fetch() support – it’s not good enough to replace the axios
backend (no cookie support), but enough to run the browser tests. So for
now, we run the “browser” tests under Node as well, at least in CI and
when you run `npm t`; but also add a browser-test.html file that can be
loaded manually in a browser (using `python -m http.server` or
equivalent – it won’t work using the file:// protocol), and tweak
browser.test.js so that the Chai import works in Node and browser.

This requires bumping the test-full CI job to Node 18; I’m generally
happy to do that, since test-most (Node 12) still checks support for
older Node versions, but at some point Node removed the
--harmony-top-level-await option, so we need to tweak the test:readme
command to retry running Node without it. (Fortunately we don’t have to
repeat the sed command, when Node complains about the invalid option it
doesn’t consume any stdin.)

I hope to eventually replace this with proper testing in a browser
again, see #23.