Add performance to global for older Node versions
Node versions between 8.5 (inclusive) and 16 (exclusive) have a
performance object, but it’s only available for import from the
node:perf_hooks internal package, not on the global scope where we want
to use it. Add a file that imports performance from the package and then
adds it to the global scope; that has an impact beyond this library, but
IMHO that’s acceptable, since Node 16 already decided that it was okay
to add the object to the global scope.

We load this file in node.js and during the unit tests. If anyone uses
axios.js directly instead of via node.js, they’re responsible for
loading the file themselves. In the Node integration tests, we load the
file via node.js; in the Browser integration tests, which we now also
run in Node (see #23), we don’t load the file, but we also don’t depend
on retry behavior, so it should be fine.

I’m not generally opposed to finding another solution for this, but for
now this fixes #1.