Use String.includes() instead of regex to test for pipes
With enough iterations (starting around ten million iterations), this
actually makes a noticeable performance difference (in Node, for a
billion iterations: half a second instead of fifteen seconds; in
Firefox, I can’t measure it as exactly, but String.includes() is still
noticeably faster). I find this surprising (why can’t they both optimize
to the same code?), but the String.includes() version isn’t much less
readable IMHO, so let’s use it if it’s a bit faster.

Note that we have to use the function generically: the array elements
might not be strings. Add some tests that cover this. (The alternative
would be String( element ).includes( '|' ), but I like that a little bit
less for some reason.)