Move action= into URL params for POST requests
Generally speaking, for a POST request, almost all params can go into
the request body (except origin=, for CORS pre-flight), and some params
must go into the body, but also it’s nicer for Wikimedia production if
params are in the request URL (then they get included in Logstash, for
example). Try to improve production logging by moving at least the
action= into the URL, which should always be safe.

I might also introduce a more general mechanism for controlling URL
parameters in future (maybe a Set of parameters that can always go in
the URL, defaulting to action, origin, formatversion, etc., and then a
Map from action to another Set of per-action URL parameters, defaulting
to empty but customizable by users and extension packages). But not
right now, and this commit should already be an improvement.