Set up JSDoc, including automatic publish on GitHub pages
Configure JSDoc. Sadly, the config file needs to be JSON – ESLint would
also understand a CommonJS file, but a .js file would be interpreted as
an ESM file due to our package.json type, and if it’s renamed to .cjs
then ESLint tries to load it as JSON again. (It looks like the future
JSDoc 5 might fix this and also add a default config path.)

The layout.tmpl is based on JSDoc’s default template, but with a comment
block added indicating where the warning about an old version will go.
And since we have to copy over the file anyways, that’s also an
opportunity to improve a few other things: change the title prefix from
“JSDoc” to “m3api”, and defer the scripts.

In the release workflow, we add a job to add the docs of the new release
to the gh-pages branch, and to also update all previous releases to
point to the latest release. (We actually do it the other way around:
it’s easier to first update all docs and only then add the new ones.)
The initial setup of the gh-pages branch will be done manually. Note
that I don’t plan to rebuild old documentation: we just keep the same
files in the Git history, updating them every time. If that doesn’t work
out, we can still change it later.

Implements #17.