Add a stable interface policy
As proposed in #13: we have a public stable interface and an internal
interface, and we’re allowed to change the internal interface between
minor versions, to leave ourselves some breathing room.

I’ll say here (but don’t consider it worth including in the README) that
I expect some level of common sense to be applied with regard to the
stable interface. For example, since the documentation “strongly
discourages” changing or removing DEFAULT_OPTIONS entries, I will not
consider changes to be incompatible just because they would
theoretically break a user who disregarded this advice.