Exclude axios 1.2.0
It has a pretty bad bug (axios/axios#5328) where it will fail to decode
responses that don’t have a Content-Length response header set; a fix is
already committed, but not yet released.

npm semver doesn’t have a syntax to exclude a specific version, but we
can simulate it by combining two ranges. (It’s not clear whether the
next axios release will be 1.2.1 or 1.3.0, but ^1.2.1 should match
either.)