Add internal retryUntil option
This option, which takes precedence over maxRetriesSeconds, is used
during retry to ensure that the time until which we retry doesn’t drift
around as we convert between absolute and relative time: we convert the
relative maxRetriesSeconds into absolute retryUntil once and then use
that time until the request is fully finished. Part of #24.