Enable eslint in combine.js
I just noticed that it was disabled there, for no good reason – I assume
that when I was first writing this code, the eslint warnings were
annoying and I wanted to be able to run `npm t` without caring about
them, and then I forgot to reenable eslint later.

Inevitably, we accumulated some eslint violations while it was disabled,
but they’re not too bad to fix.