Merge internalRequest() method into request()
Part of #1, since the retry handlers should receive a full options
object; we could pass the options as a whole into internalRequest(), or
reconstruct a new options object there, but all that just makes the
distinction between request() and internalRequest() even fuzzier than it
already is.

This makes request() a lot longer, but I think it still arguably makes
the code more readable overall. The TOKEN_PLACEHOLDER is gone, and we no
longer need to pass around half a dozen local variables nor reassemble
them into an options object for recursive calls (as we used to do for
the getToken() call). There are a few further possible cleanups that
I’ll do in separate commits; also, some of the retry-related code should
move into the DEFAULT_OPTIONS with #1.

This theoretically makes retry requests eligible for being combined with
other requests (since retry now calls request(), not internalRequest()),
but this is either vanishingly unlikely or outright impossible (even if
a new request is made at the exact same time that the first request’s
response arrives, I’m not sure if they can both happen in the same call
stack and thus be combined), so I don’t think it makes sense to add a
test for this.

All the existing tests seem to pass without modification, which makes me
more confident about this refactoring; I also manually ran the tests in
m3api-query, m3api-botpassword and m3api-oauth2:

    # in each extension package (cloned next to m3api):
    mv node_modules/m3api{,.bak} &&
    ln -s ../../m3api node_modules/ &&
    npm t &&
    unlink node_modules/m3api &&
    mv node_modules/m3api{.bak,}