Update eslint and related packages
The compat plugin now complains about our use of fetch(); silence that
with eslint-disable-next-line for now. We could apparently tell it that
we “polyfill” fetch() in our whole project, but I don’t want to do that;
I couldn’t find a way to tell it “don’t complain about fetch()
specifically” only in this file.