Drop Node 12 / 14 compatibility
Require Node 16 or later. (I’m open to even bumping this to Node 18
later, but I don’t think it’s needed at the moment.) Part of #25.

In GitHub actions, we need to keep the split between test-full and
test-most, because test:browser requires Node 18 (16 is not enough, it
still doesn’t have fetch()), but we also want to test Node 16
compatibility.