Fix upload tests in new year
I’m not sure why I included the year number in the file name in
f6e28b57a7 (probably to limit the number of file versions?), but this
made the tests fail in 2024: we expect warnings for the same content in
older versions of the same file, but not in other files. Add that.

    {"result":"Warning","warnings":{"duplicate":["M3api_test_file_2023.svg"]},"filekey":"...","sessionkey":"..."}