Override Chai version
I have no idea why `npm i` produced a package.json file that a later
`npm i` / `npm ci` would crash with, but whatever. For now, we can use
an override to allow the new version of Chai to be used, despite the
conflicting peer dependency in chai-as-promised; in the longer run, we
might have to move away from chai-as-promised to some fork of it.