Support non-200 response status codes
These are used by a few API modules, notably some WikiLambda APIs (or
Wikifunctions APIs, as they’re apparently starting to be called now). I
hope that checking for the MediaWiki-API-Error response header is enough
to distinguish between such responses and genuine internal errors that
we shouldn’t even try to decode.

Only unit-tested, because this feature is so rare that none of the APIs
using it seem suitable for an integration test against a real MediaWiki.

Fixes #30.