Add support for crossorigin parameter
This is a parameter I’m proposing for the Action API (see the linked
Gerrit change); as it affects CORS, it must be part of any preflight
request and thus be included in the URL, rather than the POST body, just
like the origin parameter. Let’s add support for it already so it can be
tested more easily. (If the parameters ends up not being merged after
all, we can remove this again.)