npm i @sinonjs/fake-timers@latest
As noted in 7de53a52da, this causes one of the tests to break; the
runMicrotasks() works around this. It makes the countTimers() assertion
partially pointless, but I can live with that (and I think it still
asserts something, just not microtasks anymore). Note that the
runMicrotasks() can’t be part of the test function itself for two
reasons: the job is (somehow) only added after the test returns (it
doesn’t exist at the end of the test function yet – you can see that by
console.log()ing clock.jobs in both places), and the failure isn’t
specific to any one test function (if you .skip() one, the failure
shifts to the next function).

I think the missing `await`s on `tickAsync()` were always wrong, but
didn’t cause problems (the problem here also turned out to be
unrelated). Let’s add them anyways.