Update eslint-plugin-chai-friendly
But not eslint itself yet, due to wikimedia/eslint-config-wikimedia#563,
and also not eslint-plugin-compat because that’s already pulled in by
eslint-config-wikimedia and I don’t want to have two copies of it in my
node_modules/.

    npm i --engine-strict eslint-plugin-chai-friendly@latest

I think this concludes the current batch of dependency updates. Still
outdated are eslint and eslint-plugin-compat, per above, and undici,
which requires Node 20+ in v7. I might drop Node 18 support once Debian
13 (Trixie) comes out (likely this summer), but certainly not before
then, so for now we’re stuck on undici 6.