Add step to release sequence
I think I sometimes remember to do this but not always, so let’s put it
in the document. Part of #23.

Also set all the list entries to start with 1. so they don’t have to be
renumbered again next time (Markdown / HTML supports this; I mostly look
at this file in Emacs but it’s not like the “wrong” numbers should
bother me ^^)