Make import in README.md example more idiomatic
Nobody outside of m3api would import it as ./node.js; the “may need to”
should really have been a “will have to”, to be honest. Let’s change the
import to the usual 'm3api', and instead rewrite the import in the
test:readme script, which is where it actually needs to be ./node.js.
Then expand the “also works in browser” explanation a bit accordingly.