Add usage recommendations to README
Hopefully this will be useful.

A quick note on OAuth owner-only clients: I tested it, and they do *not*
tag the edits (like regular clients do), so in this regard they’re equal
to bot passwords. (I think this is good, I just wasn’t sure if they
behaved that way.) Test edit:
https://test.wikipedia.org/w/index.php?title=User:Lucas_Werkmeister/sandbox&diff=prev&oldid=653493