Migrate to Wikimedia GitLab
The primary Git forge for m3api is now Wikimedia GitLab. The
documentation stays on GitHub Pages for now, so the GitHub repository
isn’t archived yet (and will be updated with each new release).

The biggest part of this work was migrating the CI from GitHub Actions
to GitLab CI, but most of the result isn’t in this repository – it’s in
a separate repository, m3api/ci-templates, and .gitlab-ci.yml here
mostly just activates the jobs from there. That said, the test-job has
limited concurrency (because the Node.js integration test edits a
testwiki page), the test-types-job is unique to this repository, and we
map the right group variables to MEDIAWIKI_USERNAME and
MEDIAWIKI_PASSWORD (this repo runs tests against production).