Add issues section to README.md
Back on GitHub, it was reasonably obvious what the correct place to
report issues was (even though it wasn’t actually ideal, as I mentioned
in T392290 – it wasn’t as easy to cross-reference other Wikimedia tasks,
and sometimes issues turned out to belong to a different repository than
where they were initially filed). With the move to GitLab, I’d like to
have all new issues on Phabricator, so let’s document that a bit more
explicitly – it also gives us an opportunity to link to several forms
for different kinds of tasks, and remind people that if they file a
security task, they should subscribe me explicitly.

As of this writing, I’m undecided whether I want to copy this section to
the README.md files of extension packages or not.