Publish documentation on doc.wikimedia.org
In m3api-query v1.1.0 I held off on updating the URL in the README until
after the release, but now that publishing the documentation was
successful there, I feel sufficiently confident that it works to
immediately update it here. Worst case, I’ll just have to push out
another patch release to revert the URL in the README again so npmjs.com
doesn’t show a broken link.